'Refactored by Sourcery'

+++ serpent/game.py
@@ -43,7 +43,7 @@ class Game(offshoot.Pluggable):
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        self.config = config.get(f"{self.__class__.__name__}Plugin", dict())
+        self.config = config.get(f"{self.__class__.__name__}Plugin", {})
 
         self.platform = kwargs.get("platform")
 
@@ -71,8 +71,8 @@ class Game(offshoot.Pluggable):
         self.api_class = None
         self.api_instance = None
 
-        self.environments = dict()
-        self.environment_data = dict()
+        self.environments = {}
+        self.environment_data = {}
 
         self.sprites = self._discover_sprites()
 
@@ -215,9 +215,12 @@ class Game(offshoot.Pluggable):
         # Look if we need to auto-append PNG to frame transformation pipeline based on given frame_handler
         png_frame_handlers = ["RECORD"]
 
-        if frame_handler in png_frame_handlers and self.frame_transformation_pipeline_string is not None:
-            if not self.frame_transformation_pipeline_string.endswith("|PNG"):
-                self.frame_transformation_pipeline_string += "|PNG"
+        if (
+            frame_handler in png_frame_handlers
+            and self.frame_transformation_pipeline_string is not None
+            and not self.frame_transformation_pipeline_string.endswith("|PNG")
+        ):
+            self.frame_transformation_pipeline_string += "|PNG"
 
         self.start_frame_grabber()
         self.redis_client.delete(config["frame_grabber"]["redis_key"])
@@ -312,7 +315,7 @@ class Game(offshoot.Pluggable):
         if self.crossbar_process is not None:
             self.stop_crossbar()
 
-        crossbar_command = f"crossbar start --config crossbar.json"
+        crossbar_command = "crossbar start --config crossbar.json"
 
         self.crossbar_process = subprocess.Popen(shlex.split(crossbar_command))
 
@@ -338,7 +341,10 @@ class Game(offshoot.Pluggable):
 
         self.redis_client.set("SERPENT:GAME", self.__class__.__name__)
 
-        input_controller_command = f"python -m serpent.wamp_components.input_controller_component"
+        input_controller_command = (
+            "python -m serpent.wamp_components.input_controller_component"
+        )
+
 
         self.input_controller_process = subprocess.Popen(shlex.split(input_controller_command))
 
@@ -359,7 +365,7 @@ class Game(offshoot.Pluggable):
 
     def _discover_sprites(self):
         plugin_path = offshoot.config["file_paths"]["plugins"]
-        sprites = dict()
+        sprites = {}
 
         sprite_path = f"{plugin_path}/{self.__class__.__name__}Plugin/files/data/sprites"
 
@@ -382,25 +388,31 @@ class Game(offshoot.Pluggable):
         return sprites
 
     def _handle_signal_frame_grabber(self, signum=15, frame=None, do_exit=True):
-        if self.frame_grabber_process is not None:
-            if self.frame_grabber_process.poll() is None:
-                self.frame_grabber_process.send_signal(signum)
+        if (
+            self.frame_grabber_process is not None
+            and self.frame_grabber_process.poll() is None
+        ):
+            self.frame_grabber_process.send_signal(signum)
 
-                if do_exit:
-                    exit()
+            if do_exit:
+                exit()
 
     def _handle_signal_crossbar(self, signum=15, frame=None, do_exit=True):
-        if self.crossbar_process is not None:
-            if self.crossbar_process.poll() is None:
-                self.crossbar_process.send_signal(signum)
+        if (
+            self.crossbar_process is not None
+            and self.crossbar_process.poll() is None
+        ):
+            self.crossbar_process.send_signal(signum)
 
-                if do_exit:
-                    exit()
+            if do_exit:
+                exit()
 
     def _handle_signal_input_controller(self, signum=15, frame=None, do_exit=True):
-        if self.input_controller_process is not None:
-            if self.input_controller_process.poll() is None:
-                self.input_controller_process.send_signal(signum)
+        if (
+            self.input_controller_process is not None
+            and self.input_controller_process.poll() is None
+        ):
+            self.input_controller_process.send_signal(signum)
 
-                if do_exit:
-                    exit()
+            if do_exit:
+                exit()
